Extract submit_job / cancel_job / download_artifacts into _submit.py
PR 3 of 3 splitting the peer_link_client/ package. Pulls the
public submit_job / cancel_job / download_artifacts flows + their
four internal helpers (_require_open_channel,
_register_submit_job_ack_future, _send_submit_job_frames,
_await_submit_job_ack) out of client.py as free functions taking
the client as first arg. The public methods stay on PeerLinkClient
as thin delegates so tests calling client.submit_job(...) /
client.cancel_job(...) / client.download_artifacts(...) (~13 sites
in test_remote_build_peer_link_client.py) keep working unchanged.

The _SUBMIT_JOB_ACK_TIMEOUT_SECONDS constant moves with the
helpers; one test patch site (4124) redirects to the new module.

client.py drops to 772 lines — under the 800-line soft cap from
#619 for the first time since this file existed. Final peer_link_client/
package shape: __init__.py 49, one_shot.py 520, _dispatch.py 426,
_submit.py 338, client.py 772, all under cap.
